+++ test/net/lewisship/group_ns.clj
@@ -1,0 +1,15 @@
+(ns net.lewisship.group-ns
+  "Group example namespace"
+  {:command-group "group"
+   :command-category "Grouped commands"}
+  (:require [net.lewisship.cli-tools :refer [defcommand]]))
+
+(defcommand echo
+  "Echo a string"
+  [:args
+   s ["TEXT" "Text to echo"]]
+  (println "echo:" s))
+
+(defcommand edit
+  "Edit a whatever"
+  [])